Add waker module in soc
