+++ crates/soc/src/lib.rs
@@ -1,4 +1,5 @@
 mod header;
 pub mod soc;
+pub(crate) mod waker;
 
 pub use crate::soc::SOC;
