+++ crates/soc/src/soc.rs
@@ -71,7 +71,7 @@ impl SOC {
     }
 
     pub fn run(&mut self) {
-        let waker = waker::create();
+        let waker = crate::waker::create();
         let mut context = Context::from_waker(&waker);
         if self.clock < 20 {
             self.run_cpu(&mut context);
@@ -87,22 +87,3 @@ pub enum Executing {
     Finished(u8),
 }
 
-pub mod waker {
-    pub use std::task::{RawWaker, RawWakerVTable, Waker};
-
-    pub fn create() -> Waker {
-        // Safety: The waker points to a vtable with functions that do nothing. Doing
-        // nothing is memory-safe.
-        unsafe { Waker::from_raw(RAW_WAKER) }
-    }
-
-    const RAW_WAKER: RawWaker = RawWaker::new(std::ptr::null(), &VTABLE);
-    const VTABLE: RawWakerVTable = RawWakerVTable::new(clone, wake, wake_by_ref, drop);
-
-    unsafe fn clone(_: *const ()) -> RawWaker {
-        RAW_WAKER
-    }
-    unsafe fn wake(_: *const ()) {}
-    unsafe fn wake_by_ref(_: *const ()) {}
-    unsafe fn drop(_: *const ()) {}
-}
